requirements.txt
Django==2.1.13
django-crispy-forms==1.9.1
django-markdownx==3.0.1
django-taggit==1.2.0
django-treenode==0.14.2